# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Auth user",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::login",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User is logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Register user",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::register",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User register successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/restore": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Restore password",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::restore",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/SendRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Email was sent successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/restore/confirm": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Register user",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::restoreConfirm",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/RestoreConfirmRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RestoreConfirmRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Password was successfully changed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user info",
                "operationId": "App\\Http\\Controllers\\API\\UserController::show",
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update user info",
                "operationId": "App\\Http\\Controllers\\API\\UserController::update",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "github": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RestoreConfirmRequest": {
                "properties": {
                    "token": {
                        "description": "Access token",
                        "type": "string"
                    },
                    "password": {
                        "description": "New password",
                        "type": "string"
                    },
                    "password_confirmation": {
                        "description": "Confirm new password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SendRequest": {
                "properties": {
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "about": {
                        "type": "string"
                    },
                    "github": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "is_finished": {
                        "type": "boolean"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "login": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "about": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "github": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "is_finished": {
                        "type": "boolean"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://test.su/auth/login",
                        "tokenUrl": "http://test.su/oauth/token",
                        "refreshUrl": "http://test.su/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}